refactor: mv lock to common
